test(python): Gather all streaming tests
