+++ py-polars/tests/unit/operations/test_aggregations.py
@@ -280,16 +280,3 @@ def test_sum_empty_and_null_set() -> None:
     df = pl.DataFrame({"a": [None, None, None], "b": [1, 1, 1]})
     assert df.select(pl.sum("a")).item() == 0.0
     assert df.groupby("b").agg(pl.sum("a"))["a"].item() == 0.0
-
-
-def test_null_sum_streaming_10455() -> None:
-    df = pl.DataFrame(
-        {
-            "x": [1] * 10,
-            "y": [None] * 10,
-        }
-    )
-    assert df.lazy().groupby("x").sum().collect(streaming=True).to_dict(False) == {
-        "x": [1],
-        "y": [0.0],
-    }
